Create LICENSE
The MIT License is a permissive open-source license that allows you to freely use, modify, and distribute the software contained in this repository for both personal and commercial purposes. You are granted extensive freedom to adapt and customize the code to suit your needs. However, it is important to note that the software is provided "as is" without any warranty or guarantee of its performance or fitness for a particular purpose.

Under the MIT License, you are required to include the original copyright notice and disclaimer in all copies of the software. You are also encouraged to acknowledge the original authors and contributors of the repository when using or sharing the code.

By using this repository, you agree to adhere to the terms and conditions of the MIT License and accept any associated risks or limitations. Please refer to the LICENSE file in this repository for the complete text of the license.